% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{check_age_groups}
\alias{check_age_groups}
\title{Stops with a clear message if age groups are not coded 1 to 18}
\usage{
check_age_groups(data, file_label)
}
\arguments{
\item{data}{The dataset to check, with an \code{age.grp} column}

\item{file_label}{The name of the file, used in the message}
}
\description{
Stops with a clear message if age groups are not coded 1 to 18
}
