% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{choose_best_model}
\alias{choose_best_model}
\title{Selects either spline or linear model based on AIC}
\usage{
choose_best_model(segmented_data, x_val, y_val)
}
\description{
Selects either spline or linear model based on AIC
}
\author{
Sean Francis
}
