% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{fit_piecewise_model_with_selection}
\alias{fit_piecewise_model_with_selection}
\title{Fit piecewise model}
\usage{
fit_piecewise_model_with_selection(data, breakpoints, x_val, y_val)
}
\description{
Fit piecewise model
}
\author{
Sean Francis
}
