% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{suppress_averages}
\alias{suppress_averages}
\title{Removes 5-year averages that are based on too few cases}
\usage{
suppress_averages(data, suppress_threshold)
}
\arguments{
\item{data}{The averages dataset}

\item{suppress_threshold}{The suppression threshold chosen by the user}
}
\description{
Removes 5-year averages that are based on too few cases
}
