% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{transform_data}
\alias{transform_data}
\title{Transform the data into an appropriate format for the application}
\usage{
transform_data(
  req_mortality_data,
  req_population_data,
  standard_pop,
  std_pop_name,
  supplied_params,
  suppress_threshold
)
}
\description{
Transform the data into an appropriate format for the application
}
\author{
Sean Francis
}
